feat: 커뮤니티 칼럼 게시판 기본 구조 설정 - community-layout-client를 page-wrapper(사이드바)와 shell(탭/필터)로 분리 - 칼럼 목록/상세/작성/수정 페이지 라우트 추가 - 커뮤니티 라우트 상수 구조화 (LIST, DETAIL, CREATE, EDIT)
