+++ src/app/(home)/community/column/[id]/page.tsx
@@ -1,0 +1,10 @@
+type Props = {
+  params: Promise<{ id: string }>;
+};
+
+export default async function ColumnDetailPage({ params }: Props) {
+  const { id } = await params;
+  const columnId = Number(id);
+
+  return <div>{columnId}</div>;
+}
