+++ src/app/(home)/community/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from 'next';
 
 import { SITE_CONFIG } from '@/config/site';
-
-import CommunityLayoutClient from './community-layout-client';
+import CommunityPageWrapper from '@/features/community/components/community-page-wrapper';
 
 export const metadata: Metadata = {
   metadataBase: new URL(SITE_CONFIG.url),
@@ -13,5 +12,5 @@ export default function CommunityLayout({
 }: {
   children: React.ReactNode;
 }) {
-  return <CommunityLayoutClient>{children}</CommunityLayoutClient>;
+  return <CommunityPageWrapper>{children}</CommunityPageWrapper>;
 }
