+++ src/app/(private)/community/column/[id]/edit/page.tsx
@@ -1,0 +1,10 @@
+type Props = {
+  params: Promise<{ id: string }>;
+};
+
+export default async function ColumnEditPage({ params }: Props) {
+  const { id } = await params;
+  const columnId = Number(id);
+
+  return <div>{columnId}</div>;
+}
