+++ src/app/(private)/community/column/new/page.tsx
@@ -1,0 +1,3 @@
+export default function ColumnNewPage() {
+  return <div>칼럼 작성</div>;
+}
